Fix signed ctype problems
Thanks to:  Eric Blake  <eblake@redhat.com>
build: avoid error message failure on 64-bit hosts
build: avoid range errors in use of ctype functions

* src/scribble.c (xscribble_get): Use correct specifier.
* src/shar.c (to_uchar): New helper, borrowed from coreutils.

+++ src/scribble.c
@@ -163,7 +163,8 @@ xscribble_get(ssize_t sz)
 {
     void * p = scribble_get(sz);
     if (p == NULL) {
-        fprintf(stderr, _("could not allocate %u bytes of scribble space"), sz);
+        fprintf(stderr, _("could not allocate %u bytes of scribble space"),
+                (unsigned int)sz);
         exit(EXIT_FAILURE);
     }
     return p;
